refactor: modularize settings and auth features by creating new components, hooks, and directory structures while removing redundant server functions.
