+++ src/lib/query-keys.ts
@@ -1,0 +1,6 @@
+export const queryKeys = {
+  auth: {
+    listUserAccounts: ['list-user-accounts'] as const,
+    listSessions: ['list-sessions'] as const,
+  },
+}
